Bug #49324: fix removal of objects where the parent object is already removed
The mapping from the deleted DN to the original DN did not work if the
parent object had also already been removed.

__dn_from_deleted_object() uses the lastKnownParent attribute of deleted objects
to build the original DN of the to be removed object. That of course
failed if the lastKnownParent contains also a DN of a removed object
(with "\nDEL: ...").